Extract bit utilies into a separate file
