+++ c/src/CMakeLists.txt
@@ -2,7 +2,7 @@ cmake_minimum_required(VERSION 2.8.7)
 
 file(GLOB librtcm_HEADERS "${PROJECT_SOURCE_DIR}/include/*.h")
 
-add_library (rtcm rtcm3_decode.c)
+add_library (rtcm rtcm3_decode.c bits.c)
 target_link_libraries(rtcm m)
 target_include_directories (rtcm PUBLIC ../include)
