+++ c/src/rtcm3_decode.c
@@ -14,86 +14,7 @@
 
 #include "rtcm3_decode.h"
 #include <math.h>
-
-/** Get bit field from buffer as an unsigned integer.
- * Unpacks `len` bits at bit position `pos` from the start of the buffer.
- * Maximum bit field length is 32 bits, i.e. `len <= 32`.
- *
- * \param buff
- * \param pos Position in buffer of start of bit field in bits.
- * \param len Length of bit field in bits.
- * \return Bit field as an unsigned value.
- */
-uint32_t getbitu(const uint8_t *buff, uint32_t pos, uint8_t len) {
-  uint32_t bits = 0;
-
-  for (uint32_t i = pos; i < pos + len; i++) {
-    bits = (bits << 1) + ((buff[i / 8] >> (7 - i % 8)) & 1u);
-  }
-
-  return bits;
-}
-
-/** Get bit field from buffer as an unsigned integer.
- * Unpacks `len` bits at bit position `pos` from the start of the buffer.
- * Maximum bit field length is 32 bits, i.e. `len <= 32`.
- *
- * \param buff
- * \param pos Position in buffer of start of bit field in bits.
- * \param len Length of bit field in bits.
- * \return Bit field as an unsigned value.
- */
-uint64_t getbitul(const uint8_t *buff, uint32_t pos, uint8_t len) {
-  uint64_t bits = 0;
-
-  for (uint32_t i = pos; i < pos + len; i++) {
-    bits = (bits << 1) + ((buff[i / 8] >> (7 - i % 8)) & 1u);
-  }
-
-  return bits;
-}
-
-/** Get bit field from buffer as a signed integer.
- * Unpacks `len` bits at bit position `pos` from the start of the buffer.
- * Maximum bit field length is 32 bits, i.e. `len <= 32`.
- *
- * This function sign extends the `len` bit field to a signed 32 bit integer.
- *
- * \param buff
- * \param pos Position in buffer of start of bit field in bits.
- * \param len Length of bit field in bits.
- * \return Bit field as a signed value.
- */
-int32_t getbits(const uint8_t *buff, uint32_t pos, uint8_t len) {
-  int32_t bits = (int32_t)getbitu(buff, pos, len);
-
-  /* Sign extend, taken from:
-   * http://graphics.stanford.edu/~seander/bithacks.html#VariableSignExtend
-   */
-  int32_t m = 1u << (len - 1);
-  return (bits ^ m) - m;
-}
-
-/** Get bit field from buffer as a signed integer.
- * Unpacks `len` bits at bit position `pos` from the start of the buffer.
- * Maximum bit field length is 64 bits, i.e. `len <= 64`.
- *
- * This function sign extends the `len` bit field to a signed 64 bit integer.
- *
- * \param buff
- * \param pos Position in buffer of start of bit field in bits.
- * \param len Length of bit field in bits.
- * \return Bit field as a signed value.
- */
-int64_t getbitsl(const uint8_t *buff, uint32_t pos, uint8_t len) {
-  int64_t bits = (int64_t)getbitul(buff, pos, len);
-
-  /* Sign extend, taken from:
-   * http://graphics.stanford.edu/~seander/bithacks.html#VariableSignExtend
-   */
-  int64_t m = ((uint64_t)1) << (len - 1);
-  return (bits ^ m) - m;
-}
+#include "bits.h"
 
 void init_data(rtcm_sat_data *sat_data) {
   for (uint8_t freq = 0; freq < NUM_FREQS; ++freq) {
